orchestrator/requirements.txt
Flask
python-dotenv
requests
jsonpickle # New: for serializing/deserializing data